Refactor Auth system
