+++ Right Posture/main.py
@@ -29,6 +29,8 @@ from PySide6.QtWidgets import QApplication
 
 os.environ["QT_FONT_DPI"] = "96" # FIX Problem for High DPI and Scale above 100%
 
+
+
 # main "1" = MainWindow() , main "0" = LoginWindow
 main = 0
 widgets = None
@@ -42,44 +44,12 @@ class LoginWindow(QMainWindow):
         self.ui.setupUi(self)
         UILoginFunctions.Function_Login_Setup(self)
         UIFunctions.LoginUiDefinitions(self)
-        Login_Buttons.defineButtons(self)
+        Login_buttons.defineButtons(self)
         self.ui.Login_Status.setText("Login")
         self.ui.Reg_Status.setText("Register")
         self.show()
 
-    def check_register(self):
-        username = self.ui.Reg_username.text()
-        password = self.ui.Reg_password.text()
-        con_Password = self.ui.Reg_password_2.text()
-        email = self.ui.Reg_email.text()
 
-        if len(username) == 0 or len(password) == 0 or len(con_Password) == 0 or len(email) == 0:
-            self.ui.Reg_Status.setText("Please fill in all inputs.")
-            self.regis_fail()
-        elif password != con_Password:
-            self.ui.Reg_Status.setText("Passwords do not match.")
-            self.regis_fail()
-        else:
-            conn = sqlite3.connect("bin/Data/Accounts.db")
-            cur = conn.cursor()
-            user_info = [username, password, email]
-            try:
-                cur.execute('INSERT INTO login_info (username, password, email) VALUES (?,?,?)', user_info)
-                conn.commit()
-                conn.close()
-                self.ui.Reg_username.clear()
-                self.ui.Reg_password.clear()
-                self.ui.Reg_password_2.clear()
-                self.ui.Reg_email.clear()
-                self.ui.Reg_Status.setStyleSheet("#Reg_Status { color: #50fa7b }")
-                self.ui.Reg_username.setStyleSheet("#Reg_username:focus { border: 2px solid rgb(91, 101, 124); }")
-                self.ui.Reg_password.setStyleSheet("#Reg_password:focus { border: 2px solid rgb(91, 101, 124); }")
-                self.ui.Reg_password_2.setStyleSheet("#Reg_password_2:focus { border: 2px solid rgb(91, 101, 124); }")
-                self.ui.Reg_email.setStyleSheet("#Reg_email:focus { border: 2px solid rgb(91, 101, 124); }")
-                self.ui.Reg_Status.setText("Register Complete !")
-            except:
-                self.ui.Reg_Status.setText("This username is already registered in the database.")
-                self.regis_fail()
 
     def check_login(self):
         username = self.ui.username.text()
@@ -134,11 +104,11 @@ class LoginWindow(QMainWindow):
 
     def enter_regis(self, event):
         if event.key() == Qt.Key_Return or event.key() == Qt.Key_Enter:
-            self.check_register()
+            Auth_system.check_register(self)
 
     # BUTTONS INTERFACE TO app_button_login
     def Login_button_Interface(self):
-        Login_Buttons.buttonClick(self)
+        Login_buttons.buttonClick(self)
 
     def mousePressEvent(self, event):
         # SET DRAG POS WINDOW
