+++ Right Posture/modules/app_auth.py
@@ -1,0 +1,39 @@
+from main import LoginWindow
+import sqlite3
+
+class Auth_system(LoginWindow):
+
+    def check_register(self):
+        username = self.ui.Reg_username.text()
+        password = self.ui.Reg_password.text()
+        con_Password = self.ui.Reg_password_2.text()
+        email = self.ui.Reg_email.text()
+
+        if len(username) == 0 or len(password) == 0 or len(con_Password) == 0 or len(email) == 0:
+            self.ui.Reg_Status.setText("Please fill in all inputs.")
+            self.regis_fail()
+        elif password != con_Password:
+            self.ui.Reg_Status.setText("Passwords do not match.")
+            self.regis_fail()
+        else:
+            conn = sqlite3.connect("bin/Data/Accounts.db")
+            cur = conn.cursor()
+            user_info = [username, password, email]
+            try:
+                cur.execute('INSERT INTO login_info (username, password, email) VALUES (?,?,?)', user_info)
+                conn.commit()
+                conn.close()
+                self.ui.Reg_username.clear()
+                self.ui.Reg_password.clear()
+                self.ui.Reg_password_2.clear()
+                self.ui.Reg_email.clear()
+                self.ui.Reg_Status.setStyleSheet("#Reg_Status { color: #50fa7b }")
+                self.ui.Reg_username.setStyleSheet("#Reg_username:focus { border: 2px solid rgb(91, 101, 124); }")
+                self.ui.Reg_password.setStyleSheet("#Reg_password:focus { border: 2px solid rgb(91, 101, 124); }")
+                self.ui.Reg_password_2.setStyleSheet("#Reg_password_2:focus { border: 2px solid rgb(91, 101, 124); }")
+                self.ui.Reg_email.setStyleSheet("#Reg_email:focus { border: 2px solid rgb(91, 101, 124); }")
+                self.ui.Reg_Status.setText("Register Complete !")
+            except:
+                self.ui.Reg_Status.setText("This username is already registered in the database.")
+                self.regis_fail()
+
