--- Right Posture/modules/ui_Custom.py
@@ -1,7 +1,0 @@
-from main import *
-
-class ui_Custom:
-
-    def setup_gui(self):
-        widgets = self.ui
-        pass
